// src/utils/master.d.ts
export namespace master {
  /** Properties of a Player. */
  interface IPlayer {
    /** Player name */
    name?: string | null;

    /** Player identifiers */
    identifiers?: string[] | null;

    /** Player endpoint */
    endpoint?: string | null;

    /** Player ping */
    ping?: number | null;

    /** Player id */
    id?: number | null;
  }
// ...
  interface IServerData {
    /** ServerData svMaxclients */
    svMaxclients?: number | null;

    /** ServerData clients */
    clients?: number | null;

    /** ServerData protocol */
    protocol?: number | null;

    /** ServerData hostname */
    hostname?: string | null;

    /** ServerData gametype */
    gametype?: string | null;

    /** ServerData mapname */
    mapname?: string | null;

    /** ServerData resources */
    resources?: string[] | null;

    /** ServerData server */
    server?: string | null;

    /** ServerData players */
    players?: master.IPlayer[] | null;

    /** ServerData iconVersion */
    iconVersion?: number | null;

    /** ServerData vars */
    vars?: { [k: string]: string } | null;

    /** ServerData enhancedHostSupport */
    enhancedHostSupport?: boolean | null;

    /** ServerData upvotePower */
    upvotePower?: number | null;

    /** ServerData burstPower */
    burstPower?: number | null;

    /** ServerData connectEndPoints */
    connectEndPoints?: string[] | null;
  }
// ...
  interface IServer {
    /** Server EndPoint */
    EndPoint?: string | null;

    /** Server Data */
    Data?: master.IServerData | null;
  }
// ...
}